fix according to review comments
