+++ modules/video/multiple_object_tracking/fairmot_dla34/config/_base_/fairmot_reader_1088x608.yml
