+++ modules/video/multiple_object_tracking/fairmot_dla34/requirements.txt
@@ -1,0 +1,3 @@
+paddlepaddle >= 2.1.0
+paddlehub >= 1.6.0
+ppdet >= 2.1.0
